requirements.txt
streamlit==0.69.1
pandas==1.1.3
google_auth_oauthlib==0.4.1
google_api_python_client==1.12.8
protobuf==3.14.0
